impl TORI primitive
>R/I (to return stack w/ index) is an additon to jonesforth's words:
it sets the returns stack element at index (top of parameter stack) to
the object at 2nd top of parameter stack.

the idea is to impl usecases like
RSP@ +1 @
or
RSP@ +1 !
where return stack elements are accessed via an address/index.

the "index" is the distance from the top of the stack, ie top element
has index=0.